# docs/api/openapi.yml
openapi: 3.0.0
info:
    title: WildDuck API
    description: WildDuck API docs
    version: 1.0.0
    contact:
        url: 'https://github.com/nodemailer/wildduck'
servers:
    - url: 'https://api.wildduck.email'
...
paths:
...
    '/users/{user}/asps/{asp}':
        delete:
            tags:
                - ApplicationPasswords
            summary: Delete an Application Password
            operationId: deleteASP
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
        get:
            tags:
                - ApplicationPasswords
            summary: Request ASP information
            operationId: getASP
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetASPResponse'
        parameters:
            - name: user
              in: path
              description: ID of the User
              required: true
              schema:
                  type: string
            - name: asp
              in: path
              description: ID of the Application Password
              required: true
              schema:
                  type: string
    '/users/{user}/asps':
        get:
            tags:
                - ApplicationPasswords
            summary: List Application Passwords
            operationId: getASPs
            parameters:
                - name: showAll
                  in: query
                  description: If not true then skips entries with a TTL set
                  schema:
                      type: boolean
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetASPsResponse'
        post:
            tags:
                - ApplicationPasswords
            summary: Create new Application Password
            operationId: createASP
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateASPRequest'
                required: true
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateASPResponse'
        parameters:
            - name: user
              in: path
              description: ID of the User
              required: true
              schema:
                  type: string
...
components:
    schemas:
...
        LastUse:
            required:
                - time
                - event
            type: object
            properties:
                time:
                    type: string
                    description: Datestring of last use or false if password has not been used
                    format: date-time
                event:
                    type: string
                    description: Event ID of the security log for the last authentication
            description: Information about last use
...
        SuccessResponse:
            required:
                - success
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
...
        GetASPResponse:
            required:
                - success
                - id
                - description
                - scopes
                - lastUse
                - created
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                id:
                    type: string
                    description: ID of the Application Password
                description:
                    type: string
                    description: Description
                scopes:
                    type: array
                    items:
                        type: string
                    description: Allowed scopes for the Application Password
                lastUse:
                    $ref: '#/components/schemas/LastUse'
                created:
                    type: string
                    description: Datestring
                    format: date-time
        GetASPsResponse:
            required:
                - success
                - results
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetASPsResult'
                    description: Event listing
        CreateASPResponse:
            required:
                - success
                - id
                - password
                - mobileconfig
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                id:
                    type: string
                    description: ID of the Application Password
                password:
                    type: string
                    description: 'Application Specific Password. Generated password is whitespace agnostic, so it could be displayed to the client as "abcd efgh ijkl mnop" instead of "abcdefghijklmnop"'
                mobileconfig:
                    type: string
                    description: Base64 encoded mobileconfig file. Generated profile file should be sent to the client with Content-Type value of application/x-apple-aspen-config.
...
        CreateASPRequest:
            required:
                - description
                - scopes
            type: object
            properties:
                description:
                    type: string
                    description: Description
                scopes:
                    type: array
                    items:
                        type: string
                    description: List of scopes this Password applies to. Special scope "*" indicates that this password can be used for any scope except "master"
                generateMobileconfig:
                    type: boolean
                    description: If true then result contains a mobileconfig formatted file with account config
                address:
                    type: string
                    description: E-mail address to be used as the account address in mobileconfig file. Must be one of the listed identity addresses of the user. Defaults to the main address of the user
                ttl:
                    type: number
                    description: 'TTL in seconds for this password. Every time password is used, TTL is reset to this value'
                sess:
                    type: string
                    description: Session identifier for the logs
                ip:
                    type: string
                    description: IP address for the logs
...
        GetASPsResult:
            required:
                - id
                - description
                - scopes
                - lastUse
                - created
            type: object
            properties:
                id:
                    type: string
                    description: ID of the Application Password
                description:
                    type: string
                    description: Description
                scopes:
                    type: array
                    items:
                        type: string
                    description: Allowed scopes for the Application Password
                lastUse:
                    $ref: '#/components/schemas/LastUse'
                created:
                    type: string
                    description: Datestring
                    format: date-time
...
    securitySchemes:
        AccessTokenAuth:
            name: X-Access-Token
            type: apiKey
            in: header
            description: |-
                If authentication is enabled in the WildDuck configuration, you will need to supply an access token in the `X-Access-Token` header.

                ```json
                {
                  "X-Access-Token": "59fc66a03e54454869460e45"
                }
                ```
security:
    - AccessTokenAuth: []